hisaab/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
psycopg2-binary 
passlib[bcrypt]
python-jose
pydantic
email-validator
python-jose
passlib[bcrypt]
python-multipart